docgen/pyproject.toml
[project]
name = "docgen"
version = "0.1.0"
description = "CLI tool that scans code repos and auto-generates documentation using RAG + LLMs"
readme = "README.md"
authors = [
    { name = "Nitram2704", email = "martinsanchez270804@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.24.1",
    "python-dotenv>=1.0",
    "pathspec==1.0.4",
    "tree-sitter==0.25.2",
    "tree-sitter-javascript==0.25.0",
    "tree-sitter-typescript==0.23.2",
    "chromadb==0.5.23",
    "filelock==3.12.2",
    "fastembed==0.4.2",
    "openai>=2.26.0",
    "tenacity>=9.1.4",
    "google-genai>=1.66.0",
    "aiolimiter>=1.1",
]

[project.scripts]
docgen = "docgen.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
]
